requirements.txt
asgiref==3.6.0
python-dateutil==2.8.2
sqlparse=0.4.3
tzdata==2022.7
psycopg2==2.9.9
six==1.16.0
Django==3.2.12
django-cors-headers==4.3.1
django-jazzmin==2.6.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
gunicorn==21.2.0
shortuuid==1.0.11
PyJWT==2.3.0
Pillow